% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workHorse.R
\name{check}
\alias{check}
\title{Check function.}
\usage{
check(tau, x)
}
\arguments{
\item{tau}{a numeric-valued scalar between 0 and 1.}

\item{x}{a numeric-valued scalar or vector.}
}
\value{
a numeric or vector of numerics of length \code{length(tau)}.
}
\description{
\code{check} computes the "check" loss function used in quantile regression.
}
