% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workHorse.R
\name{fv}
\alias{fv}
\alias{fv}
\title{Average Number of False Variables.}
\usage{
fv(model_coefficients, true_coefficients, threshold)

fv(model_coefficients, true_coefficients, threshold)
}
\arguments{
\item{model_coefficients}{a vector or matrix containing the coefficient estimates. If a matrix,
then the coefficients should vary by row and the simulations should vary by column.}

\item{true_coefficients}{a vector of indices of the true non-zero coefficients}

\item{threshold}{a scalar threshold below which coefficients should be set to zero}

\item{model_coefficients}{a vector or matrix containing the coefficient estimates. If a matrix,
then the coefficients should vary by row and the simulations should vary by column.}

\item{true_coefficients}{a vector of indices of the true non-zero coefficients}

\item{threshold}{a scalar threshold below which coefficients should be set to zero}
}
\value{
a scalar indicating the average number of zero-valued coefficients incorrectly selected to be in the model

a scalar indicating the average number of non-zero valued coefficients correctly selected to be in the model
}
\description{
Given coefficient estimates from a set of simulations, \code{fv} calculates the average number of
zero-valued coefficients that are incorrectly selected to be in the model.

Given coefficient estimates from a set of simulations, \code{tv} calculates the average number of
non-zero valued coefficients that are correctly selected to be in the model.
}
