% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workHorse.R
\name{mse}
\alias{mse}
\title{Mean Squared Error.}
\usage{
mse(model_coefficients, beta0, threshold)
}
\arguments{
\item{model_coefficients}{a vector or matrix of coefficient estimates. If a matrix,
then the coefficients should vary by row and the simulations should vary by column.}

\item{beta0}{a numeric vector containing the true coefficient values.}

\item{threshold}{a scalar threshold below which coefficients should be set to zero.}
}
\description{
Given coefficient estimates from a set of simulations, \code{mse} calculates
the mean squared error of the estimates average across all simulations.
}
