% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workHorse.R
\name{augment_data}
\alias{augment_data}
\title{Augmented Data.}
\usage{
augment_data(n_linvars, n_nonlinvars)
}
\description{
\code{augment_data} creates an augmented dataset for use in the weighted
unpenalized quantile regression as part of the algorithm described in
Sherwood and Wang (2016) for solving the penalized quantile regression problem.
}
