% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workHorse.R
\name{penplaqr}
\alias{penplaqr}
\title{Partially linear penalized quantile regression}
\usage{
penplaqr(formula, nonlinVars = NULL, tau = 0.5, lambda = NULL,
  penalty = "SCAD", penalty_deriv = NULL, a = NULL, data = NULL, subset,
  na.action, method, model = TRUE, contrasts = NULL, init_beta = NULL,
  splinesettings = NULL, epsilon = 1e-06, ...)
}
\description{
\code{penplaqr} fits a partially linear penalized quantile regression
}
