% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workHorse.R
\name{QBIC}
\alias{QBIC}
\title{High-dimensional BIC.}
\usage{
QBIC(tau, resids, p_linear, degfree)
}
\arguments{
\item{resids}{a vector of residuals from the candidate model.}

\item{p_linear}{the number of candidate linear covariates.}

\item{lambda}{a nonnegative scalar; the tuning parameter.}

\item{degree}{the degrees of freedom of the candidate model.}
}
\description{
\code{QBIC} computes the high-dimensional BIC described in Sherwood and Wang (2016).
}
